Measure input validation changes
Added sum_1 parameter to check_weights which indicates if the input weights should have a sum of 1.
Fixed docs for check_weights.

Changed measure_types_required to a list parameter which contains the measure types that are required for check_weights and check_p methods.

Removed validate_subset_sizes. If the membership/non_membership/hesitation degrees aren't equal it will be handled during the measure call instead of affecting the performance due to the checking every time.